Perfect network!
9 didn't make sense, I meant 8, because there are 8 ways to win (3 ver, 3 hor, 2 diag)
I tried bumping up to 16 when I realised that there were also another 8 ways to "lose" and it did increase the accuracy but didn't reach perfection
And then finally I realised, there might be 16 ways to win, but there are 3 possible "actions" for each of those 16!
Because if we think about first row [x x ?] if opponent is x would mean we want to play [x x o] but what if it was [x ? x] we would want to play [x o x] and so on!

Optimized over 3000 epochs and saved into model directory to reuse later! Success!
Hard work pays off <3